# Add optimized methods to compute grouped reductions