```julia
df = DataFrame(a = repeat(1:40000, outer=[20]),
               b = repeat(20000:-1:1, outer=[40]),
               c = randn(800000))
gd = groupby(df, :a)
using BenchmarkTools, Statistics

# Master
julia> @btime combine(gd, :c => sum);
  10.473 ms (439181 allocations: 17.39 MiB)

julia> @btime combine(gd, :c => sum, :b => mean);
  19.515 ms (719227 allocations: 26.24 MiB)

# This PR
julia> @btime combine(gd, :c => sum);
  1.578 ms (154 allocations: 946.95 KiB)

julia> @btime combine(gd, :c => sum, :b => mean);
  3.046 ms (186 allocations: 1.54 MiB)
```