@bkamins 2018-12-27
A quick comment (I just glanced at the code). We could add methods for counting (this is pretty common), variance, standadrd deviation, covariance. Of course this could be added later, but maybe these cases will give raise to a more general API design (I am not saying it is not general enough - but it would be good to know it would support such cases).