@bkamins 2019-01-01
I have left first comments (+ a comment on https://github.com/JuliaData/CategoricalArrays.jl/issues/178)