@matthieugomez 2019-01-03
I have to say: it is amazing how one can implement all of these fast aggregate functions in one commit. data.table / dplyr are now a few years old but I don't think they optimize `std` / `var` yet.