@nalimilan 2019-01-03
I have added commits which improve the logic and implement `var`, `std`, `first`, `last` and `length`. We don't support `count` at the moment as that function does something different (which is a bit annoying, as you need to pick an arbitrary column to get its length). When multiple pairs are passed, we now compute each function in turn, instead of computing all functions at the same time for a given group. This allows using an optimized method even if a non-optimized one is specified in another pair, makes the code simpler, and should be easier to parallelize.

> I have to say: it is amazing how one can implement all of these fast aggregate functions in one commit. data.table / dplyr are now a few years old and I don't think they optimize `std` / `var` yet.

Good point. AFAICT data.table uses a slow approach, allocating vectors for each group to compute the variance on it: https://github.com/Rdatatable/data.table/blob/b186b715837a069e06d1a49026ed501ae8a1e524/src/gsumm.c#L1151
I've mentioned this for addition to the H2O.ai benchmarks: https://github.com/h2oai/db-benchmark/issues/60