@bkamins 2019-01-03
> We don't support count at the moment as that function does something different

I see the problem, but do you see a workaround in the future?