@nalimilan 2019-01-03
I'm not sure, I haven't thought a lot about it. We could allow `() => length` or similar to avoid the need to pick an arbitrary column.