@bkamins 2019-01-04
Regarding `count` discussion:
* `count`, `any` and `all` functions could be supported for boolean vectors by `groupreduce` even given its current implementation. Right?
* then if we wanted to handle predicates with `count`, `any` and `all`, which would be useful I think then we could use signatures matching `Base.Fix1{typeof(count)}` etc. to catch and handle these cases (the current would be `by(df, :group_col, :col=>Base.Fix1(count, predicate)`, but maybe this could be shortened if `Fix1` stared being used in Base - or at least exported); this would be cleanest; but maybe it would be simpler for a user to pass a tuple `by(df, :group_col, :col=>(count, predicate)` as it should be also easy enough to match?

Of course we do not have to do this in this PR.