@nalimilan 2019-01-05
OK, I thought by `count` you mean counting rows in a group (i.e. `length`, but more intuitive). `count(pred, x)`, `any` and `all` shouldn't be too hard to support using the framework from this PR. The main issue is indeed the public API. I think we should investigate whether DataFramesMeta could automatically detect these functions (and things like `sum(skipmissing(:col))`) and transform them into something that can be optimized. Ideally that would be a generic API which can also be used directly, maybe living in SplitApplyCombine.

@andyferris @davidanthoff @piever Do you think it would be interesting/possible to move the optimized operations from this PR to a generic package like SplitApplyCombine, so that they can be shared across DataFrames, JuliaDB and Query?