@andyferris 2019-01-07
> Do you think it would be interesting/possible to move the optimized operations from this PR to a generic package like SplitApplyCombine, so that they can be shared across DataFrames, JuliaDB and Query?

Sorry Milan, I'm not terribly familiar with some of the implementation of operations on DataFrames, so I'm having a bit of trouble following along. A grouped data frame is internally sorted by grouping key, rather than nested via `Dict`s?

I guess I'm wondering precisely what you are suggesting - there is already a generic `SplitApplyCombine.groupreduce` for arbitrary iterables which is optimized not to allocate or rearrange large amounts of memory (and I would love to have `groupvar` and `groupstd` and so-on as well, and combinations of everything working well with `skipmissing`). However, in the case of `SplitApplyCombine` it uses nesting more than the traditional SQL approach so the output is a `Dict` rather than a table with two columns.

> OK, I thought by count you mean counting rows in a group

I was also torn by `groupcount` vs `grouplength`.