@nalimilan 2019-01-08
> Sorry Milan, I'm not terribly familiar with some of the implementation of operations on DataFrames, so I'm having a bit of trouble following along. A grouped data frame is internally sorted by grouping key, rather than nested via `Dict`s?

Yes, it stores a reference to a parent `DataFrame`, and information about which row goes to which group (in different forms).

> I guess I'm wondering precisely what you are suggesting - there is already a generic `SplitApplyCombine.groupreduce` for arbitrary iterables which is optimized not to allocate or rearrange large amounts of memory (and I would love to have `groupvar` and `groupstd` and so-on as well, and combinations of everything working well with `skipmissing`). However, in the case of `SplitApplyCombine` it uses nesting more than the traditional SQL approach so the output is a `Dict` rather than a table with two columns.

OK, I thought there was a function to group values according to a vector of group indices in SplitApplyCombine, but apparently that's not the case. It turns out it's often faster to compute group memberships in a first step (since you can compute hashes using SIMD and efficient memory access patterns; and with `CategoricalArray` columns group indices are already available without hashing nor dicts), and then do the actual reduction in a second step (since you can again use a very efficient loop). At least that's the approach DataFrames uses.

If you're also interested in a method taking group indices (in addition to a function computing the group, as currently), some code could be moved from DataFrames to SplitApplyCombine. It's a small part of the code, but it can be tricky, in particular when determining the initial values (especially for `minimum` and `maximum` and when skipping missing values) and the output type by progressive widening. But the fact that SplitApplyCombine returns a dict and DataFrames a vector is problematic...