@andyferris 2019-01-09
> It turns out it's often faster to compute group memberships in a first step

Right. In *SplitApplyCombine* there is `groupinds` (and `groupview`) function but it does construct a `Dict`. Also it all works with *AcceleratedArrays* where secondary indices may store precomputed groups, that can be taken advantage of for a variety of operations including `groupreduce`. (The generic `groupreduce` operation I wrote has no O(input size) allocations so supports out-of-core streaming, which is the logic behind that as the default choice, though I'm happy to revisit any implemetation details).

I imagine there are a few more operations for different permutations of splitting data. For example, just speculating, if you already have the spliting indices then we could have some `split` and `splitreduce` functions...

> But the fact that SplitApplyCombine returns a dict and DataFrames a vector is problematic...

I just see this as: we need multiple generic functions for different types of nested/flattened inputs and outputs. Some will make sense for generic iteratables, others (like the SQL flavoured group operations) depend on concepts like columns (which so far `SplitApplyCombine` isn't aware of).